Add unsupported solutions for year 6
Added two new .hc files which should produce valid solutions for the
level year 6 - rainy summer.
These are currently unsupported as they make use of the add operator,
but support for this should be coming soon.

There are two since the "easy" solution should be simpler to compile.
But the other solution should also be supported eventually, and they
should both compile down to the same instructions.