Resolved conflict in parser grammar
Conflict was caused by ambiguity in where an optional whitespace token
should go.
Specifically when parsing a variable name at the start of a line,
followed by some whitespace, eg. "foo   "
It couldn't tell whether that whitespace token should be matched as the
optional whitespace at the end of a line, or the optional whitespace
before the '=' in an assignment.
So it would guess that you were going for an assignment.
But that meant that if your line just ended there, it was a syntax error
since the next token it was expecting was an EQUALS.

Fixed this by moving all whitespace except the indent at the start of a
line to go immediately after terminals.

Enforced this by adding non-terminals for each.
Eg. EQUALS is now not used directly in its stmt rule, but is instead
used indirectly via a "equals : EQUALS optws" rule.