Fix shift/reduce conflict on equals and add operators
There was a conflict around the interpreting of expressions like:
"x = y + z", as this could have been interpreted as either x = (y + z)
or (x = y) + z by the grammar.

As it happens the parser was choosing the former, but I wanted to stop
it complaining about the conflict anyway.