Overhaul hands tracking system
Hands tracking now keeps a set of constraints on the state of the
processor's hands at each stage, rather than discrete values.
This allows more complex information about the state of the hands to be
held.
For example: following code such as: COPYFROM x; COPYTO y
the old system would only remember that the hands match y, but the new
system is capable of holding the fact that the hands match both x and y
simultaneously.

This commit also adds the new constraint ValueInHands, which allows the
tracking to know that a specific constant value (eg. 0) is in the hands
at a specific time.

This allows a pseudo-instruction LoadConstant(0) to be evaluated out
into nothing in cases when a zero is guaranteed to already be in hand,
eg. following a branch out of a JUMPZ.

This is sufficient to compile the alternative solution to year 9.