Improve multiplication by large constants
This improves code generated for multiplication of expressions by
large constant values.
This is achieved by doing smaller multiplications by each prime factor.
Each multiplication is still done by repeated addition.

This is still inefficient for values which have large prime factors.
This is most noticeable with large primes, which essentially just do
repeated addition still.
But it's also relevant for numbers with prime factors over 5.
eg. multiplication by 49 compiles a 14 instruction long sequence, but
can be done in 12 instructions.

This is enough to complete the multiplication in the year 10 solution
optimally.