Add PYTHONHASHSEED to makefile
The use of sets in the compiler leads to some non-deterministic
behaviour, where running the exact same source code through the compiler
can result in several different output programs depending on the order
in which certain sets are iterated through.

This is caused by python using a random seed when determining hashes,
which are used by sets.
This behaviour can be disabled however by using the environment variable
PYTHONHASHSEED.
Sadly, this can't be easily set during the script itself, but can at
least be set in the makefile, to ensure consistency during the unit
tests.