Add optimisations based on hand needs
Similar to the way we back-propagate variable need, and optimise based
on that, we now track and optimise on hand needs.
Eg. given the instructions `COPYFROM 0; BUMPUP 0`, we may recognise that
while the value placed in the hands by the BUMPUP instruction may go on
to be used be another instruction, the value in the hands created by
COPYFROM is never used before it is overwritten again by the BUMPUP.
Since putting a value in the hands is the sole purpose of the COPYFROM
instruction, it is rendered entirely obsolete in cases where this can be
proven to be unnecessary.